revert edition derivation refactor
